docs(docs): refresh docs and references
Align documentation, query examples, and tests with current APIs and naming.

+++ src/types/neuron.rs
@@ -37,8 +37,8 @@ pub struct NeuronInfo {
     pub validator_trust: f64,
     /// Incentive score (normalized)
     pub incentive: f64,
-    /// Emission amount
-    pub emission: f64,
+    /// Emission amount (RAO)
+    pub emission: u128,
     /// Dividends received (normalized)
     pub dividends: f64,
     /// Whether the neuron is active
